lioness/web: split setup.html and inject uboot salt into js
I'd initially hoped to use something like
  write!(file, include_str!("setup.html.template"), uuid_salt = X)
...but the '{' css and js characters get in the way, so it's easier to
split setup.html up and inject the dynamic content where it's needed.
For the uboot-provided uuid salt array, it's inserted as a const at the
top of the js. Function scope would probably be a bit nicer, but it'd
make the templating hard to follow.

Signed-off-by: David Disseldorp <ddiss@suse.de>