nanopi-neo2: firstboot detection with partitioning
When uboot detects the Lioness "firstboot" disk uuid it attempts to
create a "user" partition spanning to the end of the device. OS / "boot"
and "user" paritions are assigned random uuids via uboot, which can be
used as salt for dmcrypt when passed to Linux as bootargs.

uboot CONFIG_CMD_GPT_RENAME=y is required for "gpt read".

Signed-off-by: David Disseldorp <ddiss@suse.de>